feat : Restructure folders for database
