+++ api/database/config/config.json
@@ -1,17 +1,17 @@
 {
   "development": {
-    "username": "process.env.POSTGRES_USERNAME",
-    "password": "process.env.POSTGRES_PASSWORD",
-    "database": "process.env.POSTGRES_DATABASE",
+    "username": "root",
+    "password": null,
+    "database": "database_development",
     "host": "127.0.0.1",
     "dialect": "postgres"
   },
   "test": {
     "username": "root",
     "password": null,
-    "database": "process.env.POSTGRES_DATABASE",
+    "database": "database_test",
     "host": "127.0.0.1",
-    "dialect": "postgres"
+    "dialect": "mysql"
   },
   "production": {
     "username": "process.env.POSTGRES_USERNAME",
