+++ api/package.json
@@ -2,11 +2,11 @@
   "name": "wehungry_api",
   "version": "1.0.0",
   "description": "meal booking app ;api",
-  "main": "index.js",
+  "main": "server.js",
   "scripts": {
-    "dev": "nodemon --exec babel-node index.js",
+    "dev": "nodemon --exec babel-node server.js",
     "test": "nyc mocha --require @babel/register --exit",
-    "start": "node build/index.js",
+    "start": "node build/server.js",
     "build-babel": "babel ./ --out-dir build/ --ignore ./node_modules,./coverage,./.babelrc,./package.json,./npm-debug.log,./.travis.yml,./.eslintrc.js,./UI,./README.md",
     "clean": "rm -rf build && mkdir build",
     "build": "npm run clean && npm run build-babel",
